fix: four more consumers now agree with the verdict, and two more splits
Codex round on the authoritative pipeline; all four reproduced first.

- **P1: `recommend_release()` claimed a proven no-op it had not proven.** A
  `NO_CHANGE` verdict reached under `--contract exports` with no export
  table serialized `state: actionable`, `version_bump: none`, "No ABI or API
  changes detected" -- automation-grade advice resting on exactly the thing
  the orthogonal coverage gate was simultaneously exiting 1 about. An
  unresolved run is now `review`/`not_determined` with the reason. A
  *proven* out-of-contract exclusion stays actionable: that is a
  determination, so "no bump" really is well-founded. Same split the
  coverage exit makes, so the two cannot disagree.
- **GitHub workflow annotations** emitted `::error` for a NOT_EVALUATED
  finding, putting a red inline annotation on a passing PR. Now `::notice`,
  titled with the reason -- surfaced, not claiming to be a break.
- **The HTML report** filed the same finding under the red "Changed Symbols
  (1)" heading beside a NO_CHANGE banner and 100% compatibility, with the
  relevance nowhere on the page. It now gets its own "Not Evaluated
  (Contract)" section listing the relevance and reason, matching what
  Markdown already does.
- **`--show-only`'s `filtered_summary`** counted the unscored finding, so
  one document read `summary.breaking: 0` beside
  `filtered_summary.breaking: 1`. Now over the same evaluated subset;
  `total_changes` stays inclusive, since it counts what is displayed.

Splits, continuing the pass over files at the 2000-line hard cap:

- `buildsource/source_extractors/clang.py` 2000 -> 1653.
  `clang_nodes.py` (434) takes the single-AST-node layer: where a node came
  from, how it is spelled, what a default argument evaluates to, and what
  its body hashes to once local names are normalized. That last group is why
  the layer is worth naming -- the rename/commutativity rules exist so a
  cosmetic edit does not read as a body change, and they are used nowhere
  else. Imports nothing from `abicheck`.
- `cli_compare_helpers.py` 1999 -> 1812. `cli_compare_options.py` (250)
  takes the part that needs no engine: which flag spelling was typed, and
  the usage errors that follow. The seam is structural, not aesthetic --
  the parent is inside the baselined CLI-registration cycle, so a module
  carved out of it that reached `cli`/`cli_resolve`/`cli_dump_helpers` would
  join that cycle, which needs an ADR rather than a wider allowlist. The
  option helpers that do reach those modules stayed behind.

`dumper.py` is deliberately untouched. Tests monkeypatch 29 of its module
globals, and a moved function resolves names in its new module's namespace,
so most seams that look clean would silently disable existing patches rather
than fail. Measured, only ~174 lines can move without that hazard, and the
two largest candidate seams both cross it. Addressing the patch coupling is
its own change.
