+++ abicheck/cli_compare_helpers.py
@@ -32,7 +32,7 @@ import json
 import sys
 from collections.abc import Iterable
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, NamedTuple
+from typing import TYPE_CHECKING, Any
 
 import click
 
@@ -56,6 +56,17 @@ from .cli_audit import echo_pattern_modulations
 from .cli_compare_fold import (
     _fold_scoped_compat_into_text as _fold_scoped_compat_into_text,
     _fold_suppression_audit_into_text as _fold_suppression_audit_into_text,
+)
+from .cli_compare_options import (
+    _cli_flag,
+    _merge_cli_debug_format,
+    _NormalizedCompareOptions,
+    _param_from_cli,
+    _reject_debug_format_for_non_elf,
+    _reject_set_input_flags,
+    _resolve_debug_roots,
+    _resolve_demangle,
+    _warn_force_public_ignored,
 )
 from .cli_dump_helpers import resolve_dump_depth
 from .cli_helpers_compare import (
@@ -85,41 +96,10 @@ if TYPE_CHECKING:
     from .policy_file import PolicyFile
 
 
-def _cli_flag(name: str, value: bool) -> bool | None:
-    """Return *value* only when *name* actually came from the command line.
-
-    So a flag default (e.g. ``--scope-public-headers``'s True) doesn't mask config.
-    """
-    src = click.get_current_context().get_parameter_source(name)
-    return value if src == click.core.ParameterSource.COMMANDLINE else None
 
 
-def _param_from_cli(name: str) -> bool:
-    """True when parameter *name*'s value came from the command line (not default)."""
-    src = click.get_current_context().get_parameter_source(name)
-    return bool(src == click.core.ParameterSource.COMMANDLINE)
 
 
-def _merge_cli_debug_format(
-    debug_format_opt: str | None,
-    legacy_debug_format: str | None,
-    *,
-    legacy_from_cli: bool,
-) -> str | None:
-    """Effective *command-line* debug format across all CLI spellings (ADR-040 L2).
-
-    ``--debug-format`` (``debug_format_opt``) is the primary selector; the hidden
-    compatibility flags ``--btf``/``--ctf``/``--dwarf`` write the ``debug_format``
-    dest. Either, when typed, must beat a ``.abicheck.yml`` ``debug.format`` — so
-    fold a *command-line-sourced* legacy flag in here (the flag's own default is
-    ``None``, so ``legacy_from_cli`` distinguishes "typed" from "unset"). Returns
-    ``None`` when no format was given on the command line, letting config win.
-    """
-    if debug_format_opt is not None:
-        return debug_format_opt
-    if legacy_from_cli:
-        return legacy_debug_format
-    return None
 
 
 def _resolve_compare_config(
@@ -200,139 +180,10 @@ def _resolve_compare_config(
     return cfg_path, project_cfg, resolved_cfg, cfg_sha
 
 
-def _reject_set_input_flags(
-    exit_code_scheme: str | None,
-    reconcile_build_context: bool,
-    env_matrix_path: Path | None,
-    secondary_fmt: str | None = None,
-    used_by_apps: tuple[Path, ...] = (),
-    required_symbols: tuple[str, ...] = (),
-    diagnostic_comparison: bool = False,
-    contract_evaluation: bool = False,
-    contract_mode: str | None = None,
-    audit_suppressions: bool = False,
-    pack_paths: tuple[Path, ...] = (),
-    include_labels: dict[Path, str] | None = None,
-) -> None:
-    """Reject single-pair-only flags on a directory/package (release) compare.
-
-    The per-library fan-out has no public CLI support for these, so reject them
-    loudly rather than silently ignore them (ADR-037 D12).
-    """
-    if exit_code_scheme is not None:
-        raise click.UsageError(
-            "--exit-code-scheme is not supported for directory/package "
-            "(release) comparisons: the per-library fan-out uses the legacy "
-            "verdict scheme, or severity-aware when severity is configured in "
-            ".abicheck.yml. Compare libraries individually for explicit "
-            "scheme control."
-        )
-    if reconcile_build_context:
-        raise click.UsageError(
-            "--reconcile-build-context is not supported for directory/package "
-            "(release) comparisons; it applies to single-file / snapshot "
-            "inputs. Compare the libraries individually to use it."
-        )
-    if env_matrix_path is not None:
-        raise click.UsageError(
-            "--env-matrix is not supported for directory/package (release) "
-            "comparisons yet; it applies to single-file / snapshot inputs. "
-            "Compare the libraries individually to use it."
-        )
-    if secondary_fmt is not None:
-        raise click.UsageError(
-            "--secondary-format is not supported for directory/package "
-            "(release) comparisons yet; it applies to single-file / snapshot "
-            "inputs. Compare the libraries individually to use it."
-        )
-    if used_by_apps:
-        raise click.UsageError(
-            "--used-by is not supported for directory/package (release) "
-            "comparisons: the per-library fan-out has no per-app scoping. "
-            "Compare the specific library individually with --used-by."
-        )
-    if required_symbols:
-        raise click.UsageError(
-            "--required-symbol/--required-symbols is not supported for "
-            "directory/package (release) comparisons: the per-library "
-            "fan-out has no plugin-host-contract scoping. Compare the "
-            "specific library individually with --required-symbol."
-        )
-    if diagnostic_comparison:
-        raise click.UsageError(
-            "--diagnostic-comparison is not supported for directory/package "
-            "(release) comparisons yet: the per-library fan-out does not "
-            "wire the ADR-050 D2 comparability gate's diagnostic escape "
-            "hatch (a mismatch there still raises unhandled). Compare the "
-            "specific library individually to use it."
-        )
-    if contract_evaluation:
-        raise click.UsageError(
-            "--contract-evaluation is not supported for directory/package "
-            "(release) comparisons yet: the per-library fan-out does not "
-            "wire ADR-049 Phase 3's shadow contract evaluator. Compare the "
-            "specific library individually to use it."
-        )
-    if contract_mode is not None:
-        raise click.UsageError(
-            "--contract is not supported for directory/package (release) "
-            "comparisons yet: it selects the evidence domain for the shadow "
-            "contract evaluator, which the per-library fan-out does not wire "
-            "either. Compare the specific library individually to use it."
-        )
-    if audit_suppressions:
-        raise click.UsageError(
-            "--audit-suppressions is not supported for directory/package "
-            "(release) comparisons yet: the per-library fan-out has no "
-            "single suppression-audit result to attach. Compare the "
-            "specific library individually to use it."
-        )
-    if pack_paths:
-        raise click.UsageError(
-            "--pack is not supported for directory/package (release) "
-            "comparisons yet: the fan-out dispatches before the effective "
-            "configuration is resolved, so a pack would be accepted and then "
-            "score nothing. Compare the specific library individually to "
-            "use it."
-        )
-    if include_labels:
-        raise click.UsageError(
-            "A labeled --include (old:LABEL=PATH/new:LABEL=PATH/"
-            "both:LABEL=PATH) is not supported for directory/package "
-            "(release) comparisons yet: the per-library fan-out does not "
-            "thread ADR-050 D1's project_include_labels into its per-library "
-            "dumps, so the label would be silently dropped. Compare the "
-            "specific library individually to use it."
-        )
 
 
-class _NormalizedCompareOptions(NamedTuple):
-    collect_mode: str
-    headers: tuple[Path, ...]
-    old_headers_only: tuple[Path, ...]
-    new_headers_only: tuple[Path, ...]
-    effective_debug_format: str | None
-    demangle: bool
-    report_mode: str
-    show_impact: bool
 
 
-def _resolve_demangle(fmt: str, demangle: bool | None) -> bool:
-    """Resolve the tri-state ``--demangle`` flag against a specific format.
-
-    Default ON for the text formats whose renderer post-processes symbols
-    through ``demangle_text`` (markdown/review), OFF for machine formats
-    (json/sarif/junit) and HTML — the HTML renderer emits symbols
-    structurally and demangling its string would inject unescaped
-    ``<``/``>``/``&`` from C++ names and corrupt the markup. An explicit
-    flag always wins over the per-format default.
-
-    Shared by the primary render (:func:`_normalize_compare_options`) and
-    the ``--secondary-format`` render in :func:`run_compare`, each resolved
-    against its own format — a machine primary format paired with a text
-    secondary format (or vice versa) must not inherit the other's default.
-    """
-    return fmt in {"markdown", "review"} if demangle is None else demangle
 
 
 def _resolve_compare_collect_mode(
@@ -461,25 +312,6 @@ def _needs_inline_embed(
     )
 
 
-def _reject_debug_format_for_non_elf(
-    effective_debug_format: str | None,
-    old_fmt: str | None,
-    new_fmt: str | None,
-) -> None:
-    """Reject --debug-format / legacy --btf/--ctf/--dwarf for PE/Mach-O inputs.
-
-    They force an ELF debug format and are silently ignored by the PE/Mach-O dump
-    paths, so reject them up front (mirrors dump_cmd). JSON-snapshot / dump inputs
-    have ``*_fmt == None`` and are unaffected.
-    """
-    if effective_debug_format is None:
-        return
-    for side, bfmt in (("old", old_fmt), ("new", new_fmt)):
-        if bfmt in ("pe", "macho"):
-            raise click.BadParameter(
-                f"--debug-format {effective_debug_format} is only supported "
-                f"for ELF binaries, but the {side} input is {bfmt.upper()}."
-            )
 
 
 def _resolve_post_manifest_allowlist(
@@ -524,27 +356,8 @@ def _classify_and_reject_operands(
     return old_kind, new_kind
 
 
-def _resolve_debug_roots(
-    debug_roots: tuple[Path, ...],
-    debug_roots_old: tuple[Path, ...],
-    debug_roots_new: tuple[Path, ...],
-) -> tuple[list[Path], list[Path]]:
-    """Per-side debug roots: --debug-root old=/new= override the both-sides value."""
-    resolved_old = list(debug_roots_old) if debug_roots_old else list(debug_roots)
-    resolved_new = list(debug_roots_new) if debug_roots_new else list(debug_roots)
-    return resolved_old, resolved_new
 
 
-def _warn_force_public_ignored(
-    force_public: object, scope_public_headers: bool,
-) -> None:
-    """Warn that --public-symbol overlays need --scope-public-headers to apply."""
-    if force_public and not scope_public_headers:
-        click.echo(
-            "Warning: --public-symbol/--public-symbols-list only take effect with "
-            "--scope-public-headers; ignoring the widening overlay.",
-            err=True,
-        )
 
 
 def _app_compat_summary(result: object) -> dict[str, Any]:
